+++ examples/tools.rb
@@ -7,29 +7,26 @@ require_relative "../lib/anthropic"
 # gets API credentials from environment variable `ANTHROPIC_API_KEY`
 client = Anthropic::Client.new
 
+class GetWeather < Anthropic::BaseModel
+  required :location, String, doc: "The city and state, e.g. San Francisco, CA"
+  required :unit,
+           Anthropic::EnumOf[:celsius, :fahrenheit],
+           doc: "The unit of temperature, either 'celsius' or 'fahrenheit'",
+           nil?: true
+
+  doc "Get the current weather in a given location"
+end
+
 user_message = {
   role: "user",
   content: "What is the weather in SF?"
 }
 
-tools = [
-  {
-    name: "get_weather",
-    description: "Get the weather for a specific location",
-    input_schema: {
-      type: "object",
-      properties: {
-        location: {type: "string"}
-      }
-    }
-  }
-]
-
 message = client.messages.create(
   model: "claude-sonnet-4-20250514",
   max_tokens: 1024,
   messages: [user_message],
-  tools: tools
+  tools: [GetWeather]
 )
 
 puts "Initial response: ", message
@@ -46,6 +43,8 @@ response = client.messages.create(
   messages: [
     user_message,
     {role: message.role, content: message.content},
+    # create a new user message with type tool_result that provides the execution result of the tool
+    # ("The weather is 73f"). this simulates that we ran the weather tool and got this result.
     {
       role: "user",
       content: [
@@ -59,7 +58,7 @@ response = client.messages.create(
       ]
     }
   ],
-  tools: tools
+  tools: [GetWeather]
 )
 
 puts "Final response: ", response